requirements.txt
pyyaml
requests
mitmproxy>=7.0.2